requirements.txt
Django>=3.0.0,<5.0.0
djangorestframework
